Character DB & Logic + Adventure Viewer ENC RV
Character database now properly stores the character's information and displays them in the recycler view. Upon character generation, dice are also rolled to obtain that character's stats. Adventure viewer now has an interactive recycler view for encounters that changes the text views at the bottom of the page to fit the selected encounter :)